(Onboarding): add recovery phrase warning
